requirements.txt
aniso8601==3.0.2
appnope==0.1.0
backcall==0.1.0
certifi==2018.11.29
chardet==3.0.4
decorator==4.3.0
diff-match-patch==20181111
Django==2.1.3
django-admin-rangefilter==0.3.9
django-choices==1.6.1
django-extensions==2.1.3
django-filter==2.0.0
django-graphql-jwt==0.2.0
django-import-export==1.1.0
django-jet==1.0.8
djangorestframework==3.9.0
et-xmlfile==1.0.1
feedparser==5.2.1
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
idna==2.7
ipython==7.1.1
ipython-genutils==0.2.0
jdcal==1.4
jedi==0.13.1
Markdown==3.0.1
oauthlib==2.1.0
odfpy==1.3.6
openpyxl==2.5.10
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
promise==2.2.1
prompt-toolkit==2.0.7
ptyprocess==0.6.0
Pygments==2.2.0
PyJWT==1.6.4
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
requests==2.20.1
requests-oauthlib==1.0.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
tablib==0.12.1
traitlets==4.3.2
unicodecsv==0.14.1
urllib3==1.24.1
wcwidth==0.1.7
xlrd==1.1.0
xlwt==1.3.0
